% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStateSwitch.R
\name{dataStateSwitch}
\alias{dataStateSwitch}
\title{Alternate Observations in a Data Frame over States}
\usage{
dataStateSwitch(df, angles2, scales2, angles3, scales3)
}
\arguments{
\item{df}{A data frame returned by processNOCdata() or faultSwitch().}

\item{angles2}{Change the principal angles for State 2.}

\item{scales2}{Change the principal scales for State 2.}

\item{angles3}{Change the principal angles for State 3.}

\item{scales3}{Change the principal scales for State 3.}
}
\value{

}
\description{

}
\details{

}
