% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faultSwitch.R
\name{faultSwitch}
\alias{faultSwitch}
\title{Induce the Specified Fault on NOC Observations}
\usage{
faultSwitch(df, fault, faultStartIndex, period, shift = 2)
}
\arguments{
\item{shift}{}
}
\value{

}
\description{

}
\details{

}
