% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspProcessData.R
\name{mspProcessData}
\alias{mspProcessData}
\title{Simulate Normal or Fault Observations from a Single-State or Multi-State
Process}
\usage{
mspProcessData(fault = NULL, faultStartIndex = NULL, startTime, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{

}
\details{

}
\examples{
mspProcessData(fault = 0,
                         faultStartIndex = 8500,
                         startTime = "2016-11-27 00:00:00 CST",
                         multiState = TRUE)
}
