% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspT2Plot.R
\name{mspT2plot}
\alias{mspT2plot}
\title{T-Squared Contribution Plots}
\usage{
mspT2plot(trainData, trainLabel, trainT2, newData, newLabel, newT2, trainObs,
  var.amnt)
}
\arguments{
\item{trainData}{an xts data matrix containing the training observations}

\item{trainLabel}{Class labels for the training data as a logical (two states
only) or finite numeric (two or more states) vector or matrix column (not from
a data frame) with length equal to the number of rows in ``data." For data with
only one state, this will be a vector of 1s.}

\item{trainT2}{the Hotelling's T-squared values corresponding to the
newLabel state calculated by mspTrain using the full training data with
all variables included}

\item{newData}{an xts data matrix containing the new observation}

\item{newLabel}{the class label for the new observation}

\item{newT2}{the Hotelling's T-squared value returned by mspMonitor using
the full new observation with all variables included}

\item{trainObs}{the number of observations upon which to train the algorithm.
This will be split based on class information by a priori class membership
proportions.}

\item{var.amnt}{the energy proportion to preserve in the projection, which
dictates the number of principal components to keep}
}
\description{
Plots a variaton of the Hotelling's T-squared statistic to
visualize the contribution of each variable to a fault.
}
