% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspWarning.R
\name{mspWarning}
\alias{mspWarning}
\title{The Alarm function}
\usage{
mspWarning(mspMonitor_object, faultsToTriggerAlarm = 3)
}
\arguments{
\item{mspMonitor_object}{An xts matrix returned by the mspMonitor() function}

\item{faultsToTriggerAlarm}{Specifies how many sequential faults will cause
an alarm to trigger. Defaults to 3.}
}
\value{
An xts matrix of the same dimensions as mspMonitor_object, with a
non-NA alarm status
}
\description{
Trigger an alarm, if necessary, for incoming multivariate
process observations
}
\details{
This function and the mspMonitor function are designed to be ran
via a script within a batch. The file flow is as follows: every time
interval, run the mspMonitor function to add a flag status to each incoming
observation, and return this new xts matrix. Then, pass this same xts matrix
to the mspWarning function, which will check if the process has recorded
three or more sequential monitoring statistic flags in a row. Of note, since
these functions are expected to be repeatedly ran in real time, this
function will only check for an alarm within the last row of the xts matrix.
To check multiple rows for an alarm state, please use the mspTrain function,
which was designed to check multiple past observations.
}

