% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nineDay_xts.R
\docType{data}
\name{nineDay_clean}
\alias{nineDay_clean}
\title{Real Process Data for Training}
\format{An xts matrix of 1,155 rows and 35 features recorded over
  2017-01-17 at 00:10 to 2017-01-26 at 00:00.}
\source{
Kathryn Newhart
}
\usage{
nineDay_clean
}
\description{
Data from the SM-MBR Bioreactor system over nine days. This data
  will be used for training the mvMonitoring package.
}
\keyword{datasets}

