% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDay_xts.R
\docType{data}
\name{oneDay}
\alias{oneDay}
\title{Real Process Data for Testing}
\format{An xts matrix of 144 rows and 36 features recorded over
  2017-01-26 at 00:10 to 2017-01-27 at 00:00.}
\source{
Kathryn Newhart
}
\usage{
oneDay
}
\description{
Data from the SM-MBR Bioreactor system over 24 hours. This data
  will be used for testing the mvMonitoring package.
}
\keyword{datasets}

