% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneDay_xts.R
\docType{data}
\name{oneDay_clean}
\alias{oneDay_clean}
\title{Real Process Data for Testing}
\format{An xts matrix of 75 rows and 35 features recorded over
  2017-01-27 at 00:10 to 2017-01-27 at 12:30.}
\source{
Kathryn Newhart
}
\usage{
oneDay_clean
}
\description{
Data from the SM-MBR Bioreactor system over 12 hours. This data
  will be used for testing the mvMonitoring package.
}
\keyword{datasets}

