% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateScaleMatrix.R
\name{rotateScale3D}
\alias{rotateScale3D}
\title{Three-Dimensional Rotation and Scaling Matrix}
\usage{
rotateScale3D(rot_angles = c(0, 0, 0), scale_factors = c(1, 1, 1))
}
\arguments{
\item{rot_angles}{a list or vector containg the rotation angles in the order
following: yaw, pitch, roll. Defaults to <0,0,0>.}

\item{scale_factors}{a list or vector containing the values by which to
multiply each dimension. Defaults to <1,1,1>.}
}
\value{
A 3 * 3 projection matrix of the degree and scale changes entered.
}
\description{
Render a 3-Dimensional projection matrix given positive or
negative degree changes in yaw, pitch, and / or roll and increment or
decrement feature scales.
}
\details{
See the function commentary of "rotate_3D" for a brief explination
of how these angles behave in scatterplot3d functionality (from package
scatterplot3d).
}

