% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenDay_xts.R
\docType{data}
\name{tenDay_clean}
\alias{tenDay_clean}
\title{Real Process Data for Training}
\format{An xts matrix of 1,299 rows and 35 features recorded over
  2017-01-17 at 00:10 to 2017-01-27 at 00:00.}
\source{
Kathryn Newhart
}
\usage{
tenDay_clean
}
\description{
Data from the SM-MBR Bioreactor system over ten days. This data
  will be used for training the mvMonitoring package.
}
\keyword{datasets}

