% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataStateSwitch.R
\name{dataStateSwitch}
\alias{dataStateSwitch}
\title{Alternate Observations in a Data Frame over States}
\usage{
dataStateSwitch(df, angles2 = list(yaw = 0, pitch = 90, roll = 30),
  scales2 = c(1, 0.5, 2), angles3 = list(yaw = 90, pitch = 0, roll = -30),
  scales3 = c(0.25, 0.1, 0.75))
}
\arguments{
\item{df}{A data frame returned by processNOCdata() or faultSwitch().}

\item{angles2}{Change the principal angles for State 2.}

\item{scales2}{Change the principal scales for State 2.}

\item{angles3}{Change the principal angles for State 3.}

\item{scales3}{Change the principal scales for State 3.}
}
\value{
A data frame containing the time index, state, and feature values
  after state-specific rotation and scaling; this data frame also contains
  the other columns of df that aren't the feature values.
}
\description{
Split single-state process observations, apply multiple state
  projections, and combine these observations into a single data frame,
  arranged by process time or index.
}
\details{
This function splits a process data frame by state, and rotates and
  scales the observations from states 2 and 3 by the scales and angles
  specified in the function arguments. After state-specific rotation and
  scaling, this function combines the observations back together and orders
  them by process time index.
}
\examples{
nrml <- processNOCdata(startTime = "2016-11-27 00:00:00 CST")
dataStateSwitch(nrml)
}
