% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faultDetect.R
\name{faultDetect}
\alias{faultDetect}
\title{Process Fault Detection}
\usage{
faultDetect(threshold_object, observation, ...)
}
\arguments{
\item{threshold_object}{An object of classes "threshold" and "pca" returned
by the internal threshold() function.}

\item{observation}{A single row of an xts data matrix to compare against the
thresholds}

\item{...}{Lazy dots for additional internal arguments}
}
\value{
A named 1 * 4 matrix of the SPE statistic value ("SPE"),
SPE fault indicator ("SPE_Flag"), T2 statistic value ("T2"), and T2 fault
indicator for the single row observation passed to this function ("T2_Flag").
}
\description{
Detect if a single multivariate observation is beyond normal
parameters.
}
\details{
This function takes in a threshold object returned by the
threshold() function and a single observation as a matrix or xts row. The
function then returns a row vector of the SPE test statistics, a logical
indicator marking if this statistic is beyond the threshold, the Hotelling's
T2 statistic, and an indicator if this statistic is beyond the threshold.
Observations with monitoring statistics beyond the calculated threshold are
marked with a 1, while within-parameter observations are marked with a 0.
These threshold values are passed from the threshold() function through this
function via a returned threshold object. This object will be used in higher
function calls.

This internal function is called by faultFilter().
}

