% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspMonitor.R
\name{mspMonitor}
\alias{mspMonitor}
\title{Title}
\usage{
mspMonitor(data, labelVector, trainObs, updateFreq = cieling(0.2 * trainObs),
  faultsToTriggerAlarm = 3, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

