% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspMonitor.R
\name{mspMonitor}
\alias{mspMonitor}
\title{Real-Time Process Monitoring Function}
\usage{
mspMonitor(observations, labelVector, trainingSummary, ...)
}
\arguments{
\item{observations}{an n * p xts matrix. For real-time monitoring via batch
file, n = 1, so this must be a matrix of a single row.}

\item{labelVector}{an n * 1 integer vector of class memberships}

\item{trainingSummary}{the TrainingSpecs list returned by the mspTrain()
function. This list contains --for each class-- the SPE and T2 thresholds,
as well the projection matrix.}

\item{...}{Lazy dots for additional internal arguments}
}
\value{
an n * (p + 5) xts matrix, where the columns (p + 1):(p + 5) are the
monitoring statistics and corresponding fault flags, and an empty alarm
column
}
\description{
Monitor and flag (if necessary) incoming multivariate process
observations
}
\details{
This function is designed to be ran at specific time intervals
(every 10 seconds, 30 seconds, 1 minute, 5 minutes, 10 minutes, etc), from a
batch file which calls this function and mspWarning(). This function takes
in the specific observations to monitor and their class memberships (if any)
and returns an xts matrix of these observations concatenated with their flag
status and an empty alarm column. Users should then append these rows onto
the existing daily observations matrix. The mspWarning() function will then
take in the daily observation xts matrix with rows returned by this function
and check the monitoring statistic flag indicators to see if an alarm status
has been reached. For further details, see the mspWarning() function.
}

