% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspWarning.R
\name{mspWarning}
\alias{mspWarning}
\title{The Alarm function}
\usage{
mspWarning(mspMonitor_object, faultsToTriggerAlarm = 3)
}
\arguments{
\item{faultsToTriggerAlarm}{}
}
\description{
The Alarm function
}

