% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMonitor.R
\name{processMonitor}
\alias{processMonitor}
\title{Title}
\usage{
processMonitor(data, trainObs, updateFreq = ceiling(0.2 * trainObs),
  faultsToTriggerAlarm = 3, ...)
}
\arguments{
\item{data}{A data frame with row names as a POSIX date time label}

\item{trainObs}{How many train observations will be used}

\item{updateFreq}{How many non-flagged rows to collect before we update}

\item{faultsToTriggerAlarm}{the number of sequential faults needed to
trigger an alarm}

\item{...}{Lazy dots for additional internal arguments}
}
\description{
Apply Adaptive-Dynamic PCA to state-specific data matrices.
}
\examples{
processMonitor(MASS::mvrnorm(100, mu = c(0,0,0),
                                       Sigma = toeplitz(c(1, 0.5, 0.1))),
                         trainObs = 20, updateFreq = 5, var.amnt = 0.8)
}

