% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMonitor.R
\name{processMonitor}
\alias{processMonitor}
\title{Adaptive Process Training}
\usage{
processMonitor(data, trainObs, updateFreq = ceiling(0.2 * trainObs),
  faultsToTriggerAlarm = 3, ...)
}
\arguments{
\item{data}{An xts data matrix}

\item{trainObs}{How many train observations will be used}

\item{updateFreq}{How many non-flagged rows to collect before the function
updates}

\item{faultsToTriggerAlarm}{the number of sequential faults needed to
trigger an alarm}

\item{...}{Lazy dots for additional internal arguments}
}
\value{
A list of the following components: FaultChecks = a class specific
xts data matrix containing the SPE monitoring statistic and corresponding
logical flagging indicator, the Hotelling's T2 monitoring statitisic and
corresponding logical flagging indicator, and the Alarm indicator.
Non_Alarmed_Obs = a class specific xts data matrix of all the observations
with alarm states equal to 0. Alarms = a class-specific xts data matrix of
the features and specific alarms for Alarmed observations, where the alarm
code is as follows: 0 = no alarm, 1 = Hotelling's T2 alarm, 2 = Squared
Prediction Error alarm, and 3 = both alarms. trainSpecs = the threshold
object returned by the internal threshold() function. See this function's
help file for more details.
}
\description{
Apply Adaptive-Dynamic PCA to state-specific data matrices.
}
\details{
This function is the class-specific implementation of the Adaptive-
Dynamic PCA described in the details of the mspTrain function. See that
function's help file for further details.

This internal function is called by mspTrain(). This function calls the
faultFilter() function.
}
\examples{
data("normal_switch_xts")
# Select the data under state 1
data <- normal_switch_xts[normal_switch_xts[,1] == 1]
nTrainObs <- floor(0.2 * nrow(data))

# Remove the now unnecesary state column
featureCols <- data[,-1]

processMonitor(data = featureCols, trainObs = nTrainObs)
}

